Refactor common config
