+++ backend/src/main/java/com/csy/springbootauthbe/common/wrapper/UserDetailsServiceWrapper.java
@@ -1,4 +1,4 @@
-package com.csy.springbootauthbe.config;
+package com.csy.springbootauthbe.common.wrapper;
 
 import com.csy.springbootauthbe.user.entity.User;
 import com.csy.springbootauthbe.user.repository.UserRepository;
