+++ backend/src/main/java/com/csy/springbootauthbe/config/JWTAuthenticationFilter.java
@@ -1,7 +1,7 @@
 package com.csy.springbootauthbe.config;
 
-import com.csy.springbootauthbe.user.entity.AccountStatus;
-import com.csy.springbootauthbe.user.entity.User;
+import com.csy.springbootauthbe.common.wrapper.UserDetailsServiceWrapper;
+import com.csy.springbootauthbe.common.wrapper.UserDetailsWrapper;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -10,8 +10,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.lang.NonNull;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
